Avoid private ref-struct types
