+++ src/Intermediate/Rxmxnx.PInvoke.CString.Intermediate/CStringSequence/Utf8View.cs
@@ -263,13 +263,10 @@ public partial class CStringSequence
 			/// </summary>
 			private readonly ref struct AdvanceState
 			{
-				/// <inheritdoc cref="AdvanceState.Lengths"/>
-				private readonly ReadOnlySpan<Int32?> _span;
-
 				/// <summary>
 				/// Span of nullable integers representing the lengths of remaining segments.
 				/// </summary>
-				public ReadOnlySpan<Int32?> Lengths => this._span; // Required backing field for Mono AOT.
+				public ReadOnlySpan<Int32?> Lengths { get; }
 				/// <summary>
 				/// The length of the current element in the enumeration.
 				/// </summary>
@@ -282,7 +279,7 @@ public partial class CStringSequence
 				/// <param name="currentLength">Current item length.</param>
 				public AdvanceState(ReadOnlySpan<Int32?> lengths, Int32? currentLength)
 				{
-					this._span = lengths;
+					this.Lengths = lengths;
 					this.Current = currentLength;
 				}
 			}
